+++ app/src/main/java/com/arttttt/alwaysnotified/arch/shared/Transformer.kt
@@ -1,0 +1,3 @@
+package com.arttttt.alwaysnotified.arch.shared
+
+typealias Transformer<T, R> = (T) -> R
